pyproject.toml
[project]
name = "kis-trading"
version = "0.1.0"
description = "KIS OpenAPI 자동매매 프로그램"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # === 핵심 의존성 ===
    "aiohttp>=3.9.1",
    "aiofiles>=23.2.0",
    "websockets>=12.0",
    "asyncio-throttle>=1.0.2",
    # === 데이터 처리 ===
    "pandas>=2.1.4",
    "numpy>=1.26.2",
    # "talib-binary>=0.4.19",  # TA-Lib 바이너리 (호환성 이슈로 주석처리)
    "aiosqlite>=0.19.0",
    # === 설정 및 검증 ===
    "pydantic>=2.5.2",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.1",
    # === 스케줄링 ===
    "APScheduler>=3.10.4",
    # === 로깅 및 모니터링 ===
    "loguru>=0.7.2",
    "python-telegram-bot>=20.7",
    "pycryptodome>=3.23.0",
]

[project.optional-dependencies]
dev = [
    # === 테스팅 ===
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "faker>=21.0.0",
    "aioresponses>=0.7.4",
    
    # === 개발 도구 ===
    "black>=23.11.0",
    "isort>=5.13.2",
    "flake8>=6.1.0",
    "mypy>=1.7.1",
    
    # === 타입 힌트 ===
    "types-PyYAML>=6.0.12.12",
]

backtest = [
    # === 백테스팅 (선택사항) ===
    "backtrader>=1.9.78.123",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "isort>=5.13.2",
    "flake8>=6.1.0",
    "mypy>=1.7.1",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
strict = true
